\name{citiesData}
\docType{data}
\alias{citiesData}
\title{Population of the 999 largest German cities}
\description{Population of the 999 largest German cities as a real life example for Pareto distributed data}
\usage{data(citiesData)}
\format{List with one element}
\references{http://bevoelkerungsstatistik.de
}
\keyword{datasets}
